requirements.txt
aiogram==2.16
python-binance==1.0.15
XlsxWriter==3.0.2
